requirements.txt
pymongo
gridfs
PyMuPDF
python-docx
google-generativeai
pillow
pytesseract